server/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.4; python_version >= '3.6'
certifi==2021.5.30
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-cors-headers==3.7.0
django-influxdb-metrics==1.5.0
django-rest-passwordreset==1.1.0
django==3.2.4
djangorestframework-simplejwt==4.7.1
djangorestframework==3.12.4; python_version >= '3.5'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
influxdb==5.3.1
msgpack==1.0.2
psutil==5.8.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.1.0; python_version >= '3.6'
pymysql==1.0.2
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.17.1
python-server-metrics==0.2.1
pytz==2021.1
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
tld==0.12.6; python_version >= '2.7' and python_version < '4'
urllib3==1.26.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
